\name{BackRatesMB}
\docType{data}
\alias{BackRatesMB}

\title{Data from the Parsons et al. 2011 study: Background mutation rates}
\description{Background rates for somatic mutations used in the
  Parsons et al. 2011 medulloblastoma (MB) study.}

\usage{data(ParsonsMB11)}
\format{
  The background rates for somatic mutations used in the
  Parsons et al. MB study,
  broken down by mutation type.
  The object is a data frame which has a single row,
  with the variables representing the 25 different mutation types.
}

\references{
  Parsons DW, Li M, Zhang X, Jones S, Leary RJ, Lin J, Boca SM, Carter
  H, Samayoa J, Bettegowda C, et al.
  The genetic landscape of the childhood cancer medulloblastoma.
  \emph{Science.} DOI: 10.1126/science.1198056
  
  Parmigiani G, Lin J, Boca S, Sjoeblom T, Kinzler KW,
  Velculescu VE, Vogelstein B. Statistical methods for the analysis of
  cancer genome sequencing data. 
  \url{http://www.bepress.com/jhubiostat/paper126/}
}
\seealso{
  \code{cma.scores}, \code{cma.fdr},
  \code{cma.set.stat}, \code{cma.set.sim},
  \code{SimMethodsSims-class},
  \code{GeneCovMB}, 
  \code{GeneSampMB},
  \code{GeneAlterMB}
}

\keyword{datasets}
