\name{EntrezID2Name}
\docType{data}
\alias{EntrezID2Name}

\title{Map of gene IDs to gene names}
\description{Entrez gene identifiers used in the \code{KEGG.db} package
  are mapped to gene names.}

\usage{data(EntrezID2Name)}
\format{
  Vector having as names the Entrez gene identifiers used in the
  \code{KEGG.db} package and as entries the gene names used in the
  various data objects available.
}

\references{
  \url{ftp://ftp.genome.ad.jp/pub/kegg/pathways}
}
\seealso{
  \code{cma.set.stat}, \code{cma.set.sim}
}

\keyword{datasets}
