\name{GeneID2Name11}
\docType{data}
\alias{GeneID2Name11}

\title{Map of gene IDs to gene names for the Parsons et al. 2011
  medulloblastoma (MB) study}
\description{Gene identifiers used in the Parsons et al. 2011
  MB study are mapped to gene names.}

\usage{data(GeneID2Name11)}
\format{
  Vector having as names gene identifiers and as entries the
  gene names used in the various data objects available.
}

\seealso{
  \code{GeneAlterMB}, \code{GeneCovMB}, \code{GeneSampMB}
}

\keyword{datasets}
